improve canDelete for infocom type and crontask, check in massive actionÃ
git-svn-id: https://forge.indepnet.net/svn/glpi/trunk@9792 35612d4c-9d10-0410-837c-bf9dd1be60d7

+++ front/massiveaction.php
@@ -167,7 +167,7 @@ if (isset($_POST["itemtype"])) {
             if (class_exists($_POST["itemtype"])) {
                $item = new $_POST["itemtype"]();
                foreach ($_POST["item"] as $key => $val) {
-                  if ($val == 1) {
+                  if ($val == 1 && $item->can($key,'d')) {
                      $item->delete(array("id"=>$key));
                   }
                }
